+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * added in indicators in week/events if it is today/this week
 * added in current_week property in schedule class
 * added in events_today property in weeks class
+* added in function to event to load the officials for it
 * 
 
 ## 0.3.3